requirements.txt
beautifulsoup4==4.7.1
boto3==1.9.181
botocore==1.12.181
certifi==2019.6.16
chardet==3.0.4
docutils==0.14
idna==2.8
jmespath==0.9.4
pkg-resources==0.0.0
python-dateutil==2.8.0
python-dotenv==0.10.3
requests==2.22.0
s3transfer==0.2.1
six==1.12.0
soupsieve==1.9.2
urllib3==1.25.3
